requirements.txt
Django==3.1.5
django-model-utils==4.2.0

django-gsheets